backend/requirements.txt
fastapi
uvicorn
python-multipart
pandas
numpy
scikit-learn
joblib
torch
torchaudio
librosa
matplotlib
Pillow
tqdm
